+++ src/flows/index.ts
@@ -1,7 +1,5 @@
 export * from './drawTileStatistics';
 export * from './loadDataFromJson';
 export * from './loadTilesFromglb';
-export * from './loadUnitFromGlb';
 export * from './handleWindowEvents';
 export * from './touchTileViaRaycaster';
-export * from './phases';
