+++ src/utils/geometry.ts
@@ -2,9 +2,8 @@ import { Vector3, Vector3Tuple } from 'three';
 import { usolve } from 'mathjs';
 import { cosDeg, range } from './';
 import { Point } from '../attr';
-import { hexScreenSize } from '../flows';
 import { sinDeg } from './helpers';
-import { Tile } from '../props';
+import { Tile, GameMap } from '../props';
 class Line {
   private falsyValues = [NaN, Infinity, undefined];
 
@@ -277,12 +276,13 @@ class Hexagon {
 }
 
 const coordsToScreenPoint = (coords: Point): Vector3 => {
-  const x = hexScreenSize.x;
-  const y = hexScreenSize.y;
+  const x = GameMap.HexScreenSize.x;
+  const y = GameMap.HexScreenSize.y;
   return new Vector3(
-  y * coords.X * cosDeg(30),
-  x * cosDeg(30) * coords.Y + ((coords.X % 2) * x * sinDeg(60) / 2),
-  0);
+    y * coords.X * cosDeg(30),
+    x * cosDeg(30) * coords.Y + ((coords.X % 2) * x * sinDeg(60) / 2),
+    0
+  );
 };
 
 const highlightPath = (path: Tile[]) => {
